analyzer/models/category.py
from enum import Enum
import ast
import astor
import inspect
import sys

class CategoryEnum(str, Enum):
    AI_ML = 'Artificial intelligence and machine learning technologies and applications.'
    AUTOMATION = 'Systems automating complex tasks across various industries.'
    # ... (add all other categories here)
    UNKNOWN = 'Unclassified or unknown category.'

def show_categories():
    categories = [category.name for category in CategoryEnum if category != CategoryEnum.UNKNOWN]
    print("Available categories:")
    for category in categories:
        print(f"- {category}")

def update_category_enum(new_categories):
    global CategoryEnum
    module = ast.parse(inspect.getsource(sys.modules[__name__]))
    enum_class = next((node for node in module.body if isinstance(node, ast.ClassDef) and node.name == 'CategoryEnum'), None)
    if enum_class:
        existing_categories = {node.targets[0].id: node.value.s for node in enum_class.body if isinstance(node, ast.Assign)}
        for category in new_categories:
            if category not in existing_categories:
                new_assign = ast.Assign(targets=[ast.Name(id=category, ctx=ast.Store())], value=ast.Str(s=f'New category: {category}'))
                enum_class.body.append(new_assign)
        new_enum_dict = {name: value for name, value in existing_categories.items()}
        new_enum_dict.update({category: f'New category: {category}' for category in new_categories if category not in existing_categories})
        CategoryEnum = Enum('CategoryEnum', new_enum_dict)
        updated_source = astor.to_source(module)
        with open(__file__, 'w') as file:
            file.write(updated_source)
        print(f"CategoryEnum updated with new categories: {', '.join(new_categories)}")
    else:
        print('Error: CategoryEnum not found in the source code.')
