requirements.txt
beautifulsoup4==4.12.2
click==8.1.3
google-generativeai==0.4.0
matplotlib==3.7.1
numpy==1.24.3
orgparse==0.4.20231004
requests==2.31.0
scikit-learn==1.2.2
scipy==1.12.0
